requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic>=1.10.13,<2
casbin==1.36.2
langchain==0.2.0
langchain-openai==0.1.8
langchain-community==0.2.0
openai==1.30.0
httpx==0.27.0
sse-starlette==2.0.0
python-dotenv==1.0.1
python-multipart==0.0.9
